Fix(Promisse Error): Erro ao tratar retorno do login e logout
Sistema tentava atualizar o componente depois que o mesmo já tinha sido redirecionado para outra tela no retorno da comunicação com o backend no login e logout.
